moved `errors.rb` & `logs.rb` to `domain` level. this will help us in the future when we add errors & logs outside of the `authentication` scope
