+++ app/domain/authentication/authn_k8s/inject_client_cert.rb
@@ -1,5 +1,5 @@
 require 'command_class'
-require 'authentication/errors'
+require 'errors'
 
 module Authentication
   module AuthnK8s
