+++ app/domain/authentication/authn_oidc/authenticate_id_token/decode_and_verify_id_token.rb
@@ -1,6 +1,6 @@
 require 'uri'
 require 'openid_connect'
-require 'authentication/errors'
+require 'errors'
 
 module Authentication
   module AuthnOidc
