+++ app/domain/authentication/authn_oidc/authenticate_id_token/fetch_provider_certificate.rb
@@ -1,6 +1,6 @@
 require 'uri'
 require 'json'
-require 'authentication/errors'
+require 'errors'
 
 module Authentication
   module AuthnOidc
