+++ app/domain/authentication/validate_origin.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require 'authentication/errors'
+require 'errors'
 
 module Authentication
