+++ app/domain/errors.rb
