+++ app/domain/logs.rb
